feat(cashflow): add CashflowCategoryDrawer component for category breakdown; refactor CashflowKpiCarousel to utilize the new drawer; enhance DrawerContent with customizable minHeight prop
